requirements.txt
django
djangorestframework
djangorestframework-gis
psycopg2